requirements.txt
streamlit==1.37.1
pandas==2.2.2
openai>=1.40.0
python-dotenv>=1.0.1
